More detail on cache timeout implementation
I was quite confused how the DB cache was able to cull expired cache entries, asked on IRC and received this response. I think it'd be good to include this because others like me might be concerned that the timeout totally doesn't work on some backends. It does, of course (django is wonderful), and the implementation is clever enough to note.